+++ CHANGELOG.md
@@ -1,7 +1,7 @@
-### NEXT_VERSION_TYPE=MINOR
-### NEXT_VERSION_DESCRIPTION_BEGIN
+## [6.9.0](https://bitbucket.yamoney.ru/projects/BACKEND-GRADLE-PLUGINS/repos/gradle-project-plugin/pull-requests/15) (21-02-2021)
+
 * Поднята версия artifact-release-plugin до 3.8.1
-### NEXT_VERSION_DESCRIPTION_END
+
 ## [6.8.0](https://bitbucket.yamoney.ru/projects/BACKEND-GRADLE-PLUGINS/repos/gradle-project-plugin/pull-requests/14) (12-02-2021)
 
 * Подключение java-module-plugin и artifact-dependencies-plugin заменено на подключение